Adicionando uma base page para formar as outras pages tiver que mudar o codigo do menu para usar a base_page como classe pai e aproveitar sua herança
