+++ src/config.py
@@ -11,11 +11,7 @@ FONT_POPINS = os.path.join(PASTA_ANTERIOR, "assets", "fonts", "Poppins-SemiBold.
 
 LOGO_PATH = os.path.join(PASTA_ANTERIOR, "assets", "images", "logo.png")
 
-BOTAO_LOGIN = os.path.join(PASTA_ANTERIOR, "assets", "images", "botao_login.png")
-BOTAO_LOGIN_HOVER = os.path.join(PASTA_ANTERIOR, "assets", "images", "botao_login_hover.png")
-
 EYE_PATH = os.path.join(PASTA_ANTERIOR, "assets", "images", "eye.png")
 CLOSE_EYE_PATH = os.path.join(PASTA_ANTERIOR, "assets", "images", "invisible.png")
 
 BTN_NORMAL = os.path.join(PASTA_ANTERIOR, "assets", "images", "btn_normal.png")
-BTN_HOVER = os.path.join(PASTA_ANTERIOR, "assets", "images", "btn_hover.png")
